requirements.txt
fastapi
uvicorn
google-ads==24.0.0
python-multipart
